weather_elt_streamlit/requirements.txt
streamlit
pandas
matplotlib
snowflake-connector-python
sqlalchemy
snowflake-sqlalchemy
pyarrow<19.0.0
python-dotenv